Loops, Decisions, and Shell Scripting/assignment/assignment-3-template.py
#!/usr/bin/env python
"""TODO: Say what the code does

TODO: Elaborate on what the code does
"""

import argparse


def get_args():
    """Return parsed command-line arguments."""

    parser = argparse.ArgumentParser(
        description="TODO: Say what the code does)",
        formatter_class=argparse.ArgumentDefaultsHelpFormatter)

    # get a list of text files to process
    parser.add_argument('file_list',  # variable to access this data later: args.file_list
                        metavar='FILE', # shorthand to represent the input value
                        help='Provide file name to process. For multiple files, separate their names with spaces.', # message to the user, it goes into the help menu
                        type=str, 
                        required=True,
                        nargs="+" # will combine multiple textfile inputs into a list 
                        )

    return(parser.parse_args())


def identify_sequence(sequence):
    """TODO: Say what the function does"""

    # TODO store nucleic acid or amino acid in the variable sequence_type

    return(sequence_type)


if __name__ == "__main__":
    # TODO: get the arguments
    # TODO: loop through args.file_list to:
    #    1) Open each file
    #    2) Identify the sequence within the file (hint: call your indentify_sequence() function)
    #    3) Print the filename, and its identity to the Terminal (don't print the sequence itself)
