Loops, Decisions, and Shell Scripting/assignment/tests/test_assignment-3-template.py
#!user/bin/env python3
"""Test behavior of identify_sequence_refactored.py"""

from identify_sequence_refactored import identify_sequence # import the function to test (make sure you rename your script to "indentify_sequence_refactored.py")

def test_dna_sequence():
    """Identify a DNA sequence"""
    assert identify_sequence("ATGATGA") == "nucleic acid", "expect ATGATGA identifies as nucleic acid"

def test_dna_lowercase_sequence():
    """Identify a DNA sequence in lowercase"""
    # TODO write assert statement

def test_rna_sequence():
    """Identify an RNA sequence"""
    # TODO write assert statement

def test_aminoacid_sequence():
    """Identify an amino acid sequence"""
    # TODO write assert statement

# Note: we expect this test to fail right now, but perhaps we can refactor the code again to make it pass in the future!
# For now, let's write the test to have that reminder.
def test_nonsequence():
    """TODO: write what you think should happen with a non-sequence"""
    assert identify_sequence("ZZXy43") == "TODO: put something here", "TODO: write the message"
